Move LargePageTitle composable to same package as other composables
